tried to make remove Tab work correct, still no luck

+++ NovusNodo/wwwroot/node_framework/node_framework.js
@@ -43,6 +43,10 @@ export function createCanvas(id, reference) {
  */
 export function removeCanvasTab(id) {
     console.log("Removing Canvas Tab " + id, canvasTabs[id]);
+    
+    if(canvasTabs[id] == null)
+        return;
+        
     canvasTabs[id].delete();
     delete canvasTabs[id];
 }
